#34: Refactored contents of GATCGCardDataCollectionRenderer.render into GATCGCardDataTab
